Test angela

+++ src/main/java/com/example/demo/service/serviceImpl/ArticulosAdquiridosServiceImpl.java
@@ -43,7 +43,6 @@ public class ArticulosAdquiridosServiceImpl implements ArticulosAdquiridosServic
     @Override
     public String actualizar(ArticulosAdquiridos articulosAdquiridos) throws Exception {
         validacionesActualizar(articulosAdquiridos);
-
         articulosAdquiridosDAO.save(articulosAdquiridos);
         return "Se actualizo correctamente el artículo adquirido.";
     }
@@ -54,7 +53,7 @@ public class ArticulosAdquiridosServiceImpl implements ArticulosAdquiridosServic
             throw new Exception("Se debe ingresar el id del artículo adquirido.");
         }
         if(!articulosAdquiridosDAO.existsById(idArticuloA)){
-            throw new Exception("El artrículo adquirido con id: " + idArticuloA + " no existe.");
+            throw new Exception("El artrículo adquirido con ese id no existe.");
         }
         articulosAdquiridosDAO.deleteById(idArticuloA);
         return "El artículo adquirido se ha eliminado exitosamente.";
@@ -63,12 +62,15 @@ public class ArticulosAdquiridosServiceImpl implements ArticulosAdquiridosServic
     @Override
     public String eliminarPorIds(Long idEstudiante, Long idArticulo) throws Exception {
         if(!estudianteDAO.existsById(idEstudiante)){
-            throw new Exception("El estudiante con ese id no existe en la bd.");
+            throw new Exception("El estudiante con ese id no existe.");
         }
         if(!articulosDAO.existsById(idArticulo)){
-            throw new Exception("El articulo con ese id no existe en la bd.");
+            throw new Exception("El articulo con ese id no existe.");
         }
         ArticulosAdquiridos articulosAdquiridos = articulosAdquiridosDAO.findByIdArticuloAndIdEstudiante(idEstudiante, idArticulo);
+        if(articulosAdquiridos == null){
+            throw new Exception("No se encontró ningún registro correspondiente en la base de datos.");
+        }
         articulosAdquiridosDAO.deleteById(articulosAdquiridos.getIdArticuloAdquirido());
         return "Se elimino el articulo correctamente";
     }
@@ -79,7 +81,7 @@ public class ArticulosAdquiridosServiceImpl implements ArticulosAdquiridosServic
             throw new Exception("Se debe ingresar el id del artículo adquirido.");
         }
         if(!articulosAdquiridosDAO.existsById(idArticuloA)){
-            throw new Exception("El artículo adquirido con id: " + idArticuloA + " no existe.");
+            throw new Exception("El artículo adquirido con ese id no existe.");
         }
         return articulosAdquiridosDAO.findById(idArticuloA).get();
     }
@@ -98,17 +100,17 @@ public class ArticulosAdquiridosServiceImpl implements ArticulosAdquiridosServic
         Estudiante estudiante = estudianteDAO.findById(idEstudiante).get();
         Articulos articulo = articulosDAO.findById(idArticulo).get();
 
-        if(articulo.getIdArticulo() == null){
+        if(!articulosDAO.existsById(idArticulo)){
             throw new Exception("El articulo con ese id no existe.");
         }
-        if(estudiante.getIdEstudiante() == null){
+        if(!estudianteDAO.existsById(idEstudiante)){
             throw new Exception("El estudiante con ese id no existe.");
         }
         Double monedasEstudiante = Double.valueOf(estudiante.getMonedasObtenidas());
         Double precio = articulo.getPrecio();
 
         if(monedasEstudiante < precio){
-            throw new Exception("No tiene monedas suficientes para comprae este articulo.");
+            throw new Exception("No tienes suficientes monedas para comprar este articulo.");
         }
         monedasEstudiante = monedasEstudiante - precio;
         ArticulosAdquiridos articulosAdquiridos = new ArticulosAdquiridos();
@@ -131,7 +133,6 @@ public class ArticulosAdquiridosServiceImpl implements ArticulosAdquiridosServic
         if(!estudianteDAO.findById(articulosAdquiridos.getEstudiante().getIdEstudiante()).isPresent()){
             throw new Exception("Debe ingresar el id de un estudiante que este registrado.");
         }
-
         if(articulosAdquiridos.getArticulos().getIdArticulo() == null || articulosAdquiridos.getArticulos().getIdArticulo() <0){
             throw new Exception("Debe ingresar un artículo válido.");
         }
@@ -144,11 +145,10 @@ public class ArticulosAdquiridosServiceImpl implements ArticulosAdquiridosServic
                 || Validaciones.isStringLenght(articulosAdquiridos.getUsuarioCreador(),50)){
             throw new Exception("Se debe ingresar un usuario creador del artículo válido.");
         }
-        if(articulosAdquiridos.getFechaCreacion() == null || articulosAdquiridos.getFechaCreacion().toString().equals("")){
+        if(articulosAdquiridos.getFechaCreacion() == null){
             throw new Exception("Se debe ingresar una fecha válida.");
         }
         Date fechaActual = new Date();
-
         if(articulosAdquiridos.getFechaCreacion().compareTo(fechaActual)>0){
             throw new Exception("No puede ingresar una fecha que aun no ha sucedido.");
         }
@@ -157,19 +157,16 @@ public class ArticulosAdquiridosServiceImpl implements ArticulosAdquiridosServic
         if(articulosAdquiridos.getIdArticuloAdquirido() == null){
             throw new Exception("Debe ingresar el id del artículo que desea actualizar.");
         }
+        if(articulosAdquiridos.getIdArticuloAdquirido()<0){
+            throw new Exception("Debe ingresar el id del artículo adquirido válido.");
+        }
         if(!articulosAdquiridosDAO.existsById(articulosAdquiridos.getIdArticuloAdquirido())){
             throw new Exception("No se encontró un artículo adquirido con ese id.");
         }
-        if(!articulosDAO.existsById(articulosAdquiridos.getArticulos().getIdArticulo())){
-            throw new Exception("No se encontró un artículo con ese id.");
-        }
-        if(!estudianteDAO.existsById(articulosAdquiridos.getEstudiante().getIdEstudiante())){
-            throw new Exception("No se encontró un estudiante con ese id.");
-        }
-        if(articulosAdquiridos.getArticulos().getIdArticulo().toString().equals("")){
+        if(articulosAdquiridos.getArticulos().getIdArticulo() == null){
             throw new Exception("Debe ingresar un id de artículo.");
         }
-        if(articulosAdquiridos.getEstudiante().getIdEstudiante().toString().equals("")){
+        if(articulosAdquiridos.getEstudiante().getIdEstudiante() == null){
             throw new Exception("Debe ingresar un id de un estudiante.");
         }
         if(articulosAdquiridos.getArticulos().getIdArticulo()<0){
@@ -178,12 +175,11 @@ public class ArticulosAdquiridosServiceImpl implements ArticulosAdquiridosServic
         if(articulosAdquiridos.getEstudiante().getIdEstudiante()<0){
             throw new Exception("Debe ingresar un id de estudiante válido.");
         }
-        if(!articulosDAO.findById(articulosAdquiridos.getArticulos().getIdArticulo()).isPresent()){
-            throw new Exception("Debe ingresar un id de un articulo que exista.");
+        if(!articulosDAO.existsById(articulosAdquiridos.getArticulos().getIdArticulo())){
+            throw new Exception("No se encontró un artículo con ese id.");
         }
-        if(!estudianteDAO.findById(articulosAdquiridos.getEstudiante().getIdEstudiante()).isPresent()){
-
-            throw new Exception("Debe ingresar un id de un estudiante que exista.");
+        if(!estudianteDAO.existsById(articulosAdquiridos.getEstudiante().getIdEstudiante())){
+            throw new Exception("No se encontró un estudiante con ese id.");
         }
         Date fechaActual = new Date();
         if(articulosAdquiridos.getUsuarioModificador() == null || articulosAdquiridos.getUsuarioModificador().equals("")){
@@ -192,7 +188,7 @@ public class ArticulosAdquiridosServiceImpl implements ArticulosAdquiridosServic
         if(Validaciones.isStringLenght(articulosAdquiridos.getUsuarioModificador(),50)){
             throw new Exception("El nombre de usuario modifcador es muy largo, solo puede contener 50 caracteres.");
         }
-        if(articulosAdquiridos.getFechaModificacion() == null || articulosAdquiridos.getFechaModificacion().toString().equals("")){
+        if(articulosAdquiridos.getFechaModificacion() == null){
             throw new Exception("Debe ingresar una fecha de modificación.");
         }
         if(articulosAdquiridos.getFechaModificacion().compareTo(fechaActual)>0){
